+++ Block/Adminhtml/System/Config/Color.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Monogo\PagespeedAnalysis\Block\Adminhtml\System\Config;
 
-use Magento\Backend\Block\Template\Context;
 use Magento\Config\Block\System\Config\Form\Field;
 use Magento\Framework\Data\Form\Element\AbstractElement;
 
@@ -16,31 +17,18 @@ use Magento\Framework\Data\Form\Element\AbstractElement;
 class Color extends Field
 {
     /**
-     * Color constructor.
-     *
-     * @param Context $context Context
-     * @param array   $data    Data
-     */
-    public function __construct(
-        Context $context,
-        array $data = []
-    ) {
-        parent::__construct($context, $data);
-    }
-
-    /**
      * Get element html
      *
      * @param AbstractElement $element AbstractElement
      *
      * @return string
      */
-    protected function _getElementHtml(AbstractElement $element)
+    protected function _getElementHtml(AbstractElement $element) : string
     {
         $html = $element->getElementHtml();
         $value = $element->getData('value');
 
-        $html .= '<script type="text/javascript">
+        return $html . '<script type="text/javascript">
             require(["jquery","jquery/colorpicker/js/colorpicker"], function ($) {
                 $(document).ready(function () {
                 	var $el = $("#' . $element->getHtmlId() . '");
@@ -55,6 +43,5 @@ class Color extends Field
             	});
         	});
 	        </script>';
-        return $html;
     }
 }