+++ Model/System/Config/Strategy.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Monogo\PagespeedAnalysis\Model\System\Config;
 
 use Magento\Framework\Data\OptionSourceInterface;