+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG
 
 <a name="1.0.1"></a>
+## 1.2.0 - 2020-07-29
+- PHP 7.4 compatibility
+- Magento 2.4 Compatibility
+
+<a name="1.2.0"></a>
+## 1.2.0 - 2020-07-29
+- PHP 7.4 Compatibility
+- Magento 2.4 Compatibility
+
+<a name="1.1.0"></a>
 ## 1.1.0 - 2020-06-25
 Change time-to-first-byte to server-response-time